Remove redundant BubbleSort folder
